requirements.txt
functions-framework==3.*
firebase-functions==0.1.0
firebase-admin==6.*
pyswisseph==2.10.3.2
